# Valley floor filter performance improvement