Improve performance of the `ValleyFloorFilter`.

Currently improves performance by a factor of slightly less than 2x:

## Baseline (main)
```bash
time python -mcProfile -o test.prof scripts/benchmark.py 
/home/thlun8736/Documents/work/pyaro/src/pyaro/plugins.py:27: RuntimeWarning: Engine 'actrisebas' loading failed:
No module named 'tomllib'
  warnings.warn(f"Engine {name!r} loading failed:\n{ex}", RuntimeWarning)
59564

real    9m5,684s
user    7m26,001s
sys     0m6,409s
```

## This branch
```bash
time python -mcProfile -o test.prof scripts/benchmark.py 
/home/thlun8736/Documents/work/pyaro/src/pyaro/plugins.py:27: RuntimeWarning: Engine 'actrisebas' loading failed:
No module named 'tomllib'
  warnings.warn(f"Engine {name!r} loading failed:\n{ex}", RuntimeWarning)
59502

real    5m29,457s
user    4m6,028s
sys     0m5,225s
```

I haven't quite figured out how to vectorize the inner loop (in _calculate_relative_altitude) properly while maintaining the subset calculation, so that still uses a loop.