@thorbjoernl 2025-04-29
From looking around I found no easy way of subsetting the topo file in a vectorized manner (the examples I found all go via lists, and then concatenating) or assume regular patterns (eg. stride), but I now use numpy arrays and masking in the inner loop, and it further speeds up the calculation.

Comparison `main` vs this branch:

![image](https://github.com/user-attachments/assets/00cbdb72-d1c8-4db4-b5f9-93e4b3de7500)

![image](https://github.com/user-attachments/assets/74f68231-0598-4ebd-88fc-3951531ebacb)