@heikoklein 2025-04-29
After modifying fileformat of GTOPO files from netcdf3 to compressed/chunked netcdf4, the new `Total Nanoseconds Cost: ~14 000 000 000 ns`